refactor(broker): adapt JobStreamer API
Replaces old `GatewayStreamer` adapters with the new `JobStreamer` API.
