+++ broker/src/main/java/io/camunda/zeebe/broker/partitioning/PartitionManagerImpl.java
@@ -23,13 +23,11 @@ import io.camunda.zeebe.broker.system.monitoring.DiskSpaceUsageMonitor;
 import io.camunda.zeebe.broker.system.partitions.PartitionHealthBroadcaster;
 import io.camunda.zeebe.broker.system.partitions.ZeebePartition;
 import io.camunda.zeebe.broker.transport.commandapi.CommandApiService;
+import io.camunda.zeebe.engine.processing.streamprocessor.JobStreamer;
 import io.camunda.zeebe.protocol.impl.encoding.BrokerInfo;
 import io.camunda.zeebe.scheduler.ActorSchedulingService;
 import io.camunda.zeebe.scheduler.ConcurrencyControl;
 import io.camunda.zeebe.snapshots.impl.FileBasedSnapshotStoreFactory;
-import io.camunda.zeebe.stream.api.ActivatedJob;
-import io.camunda.zeebe.stream.api.GatewayStreamer;
-import io.camunda.zeebe.stream.api.JobActivationProperties;
 import io.camunda.zeebe.transport.impl.AtomixServerTransport;
 import io.camunda.zeebe.util.health.HealthStatus;
 import java.util.ArrayList;
@@ -62,7 +60,7 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
   private final CommandApiService commandApiService;
   private final ExporterRepository exporterRepository;
   private final AtomixServerTransport gatewayBrokerTransport;
-  private final GatewayStreamer<JobActivationProperties, ActivatedJob> jobStreamer;
+  private final JobStreamer jobStreamer;
 
   public PartitionManagerImpl(
       final ActorSchedulingService actorSchedulingService,
@@ -75,7 +73,7 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
       final CommandApiService commandApiService,
       final ExporterRepository exporterRepository,
       final AtomixServerTransport gatewayBrokerTransport,
-      final GatewayStreamer<JobActivationProperties, ActivatedJob> jobStreamer) {
+      final JobStreamer jobStreamer) {
     this.gatewayBrokerTransport = gatewayBrokerTransport;
 
     snapshotStoreFactory =
